refactor: 상품 정보 수정 조회 실패 시 예외 반환 책임 위치 변경
기존: 변경 정보가 없다면 DAO에서 예외 발생
변경 후: 변경 정보가 없다면 서비스에서 예외 발생
